Adjusted post endpoint, added like endpoint, not working
